+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 `http-directory` uses [PVP Versioning](https://pvp.haskell.org).
 
-## 0.1.4 (2019-06-xx)
+## 0.1.5 (2019-06-xx)
+- also filter ".." and "#" from httpDirectory
+
+## 0.1.4 (2019-06-07)
 - add httpRawDirectory
 - httpDirectory now filters out absolutes hrefs and sort links
 - add httpDirectory' and httpRedirect' variants with own Manager